refactor: Simplify the mode selection
The selection of which mode was shown was pretty complex. This is now
simplified.

First, when the mode is `auto`, the requested mode is adapted depending
on the first use counter.
Then, it is determined if the requested mode can be fulfilled or if a
fallback mode should be selected.

After this simple process, the mode is determined.
Based on #47 from BurningKarl